Move infer::canonical datatypes to infer::types.
